requirements.txt
apache-airflow-providers-google==10.16.0
google-cloud-bigquery==3.25.0
google-cloud-storage==2.17.0
pandas==2.2.2
pyarrow==16.0.0
scikit-learn==1.5.1
